+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `project list` can print the projects as JSON and CSV.
+- `project list` command default print format shows the client name and id
+
 ## [v0.15.1] - 2020-09-30
 
 ### Fixed